processor/requirements.txt


azure-monitor-opentelemetry

azure-functions==1.20.0

azure-storage-blob==12.20.0

requests

